@@ -71,10 +71,11 @@
     "less-plugin-clean-css": "^1.5.1",
     "open-browser-webpack-plugin": "^0.0.5",
     "postcss-loader": "^1.3.3",
-    "react": "^15.4.2",
-    "react-addons-css-transition-group": "^15.4.2",
+    "prop-types": "^15.5.6",
+    "react": "^15.5.3",
+    "react-addons-css-transition-group": "^15.5.2",
     "react-amap": "^0.2.4-0",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.5.3",
     "react-hot-loader": "3.0.0-beta.6",
     "react-router": "^3.0.0",
     "rimraf": "^2.6.1",
