@@ -1,7 +1,8 @@
 /**
  * Created by Freeman on 2017/3/30.
  */
-import React, { PropTypes } from 'react'
+import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from 'classnames'
 const Loading = ({size,prefixCls, className}) => {
 
